refactor: move more functions out of SchemaGenerator
